+++ src/pages/Pedidos.jsx
@@ -1,7 +1,7 @@
 import React, { useState } from 'react'
-import { CardDrink } from '../components/CardDrink'
+import { Pedido } from '../components/Pedido'
 import { bebidas } from '../../public/API/bebidas'
-export function Bebidas({ items, setItems }) {
+export function Pedidos({ items, setItems }) {
 
   const [drinks, setDrinks] = useState(bebidas)
 
@@ -9,7 +9,7 @@ export function Bebidas({ items, setItems }) {
     <div className='flex w-full min-h-screen bg-[#ECECEC]'>
       <div className="grid grid-cols-2 md:grid-cols-4 w-[100%] h-[100%] gap-2 md:gap-4 ms-2 md:ms-6 mt-5">
         {drinks.map(drink =>
-          <CardDrink
+          <Pedido
             drink={drink}
             key={drink.id}
             items={items}
